+++ index.html
@@ -231,6 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script scr="js/main.js"></script>
+  <script scr="js/main.js" type = "module" ></script>
 </body>
 </html>
