+++ js/consts.js
@@ -1,0 +1,42 @@
+const NAMES = [
+  'Андрей',
+  'Мишель',
+  'Мария',
+  'Алиса',
+  'Виктор',
+];
+
+const DESCRIPTIONS=[
+  'На отдыхе',
+  'Отличное место',
+  'Люблю котов',
+  'А в ответ тишина.'
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const AMOUNT_OF_ELEMENTS = 25;
+
+const countLike = {
+  MIN: 15,
+  MAX: 200
+};
+
+const commentsCount = {
+  MIN: 1,
+  MAX: 6
+}
+
+export {NAMES};
+export {DESCRIPTIONS};
+export {MESSAGE};
+export {AMOUNT_OF_ELEMENTS};
+export {countLike};
+export {commentsCount};
